// src/types/formTemplate.ts
export interface IFormTemplate {
  usr_name: string,
  usr_email: string,
  usr_password: string,
  confirmPassword: string,
  establishmentName: string,
  establishmentPhone: string,
  establishmentCep: string,
  establishmentAddress: string,
  establishmentCity: string,
};


export interface IRegisterData {
  name: string
  email: string
  password: string
  establishmentName: string
  establishmentPhone: string
  establishmentCep: string
  establishmentAddress: string
  establishmentCity: string
  confirmPassword: string
}